Move base subculture classes into module
